% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_smm.R
\name{predict.splinemixmeta}
\alias{predict.splinemixmeta}
\title{Make predictions from a fitted \code{splinemixmeta} model}
\usage{
\method{predict}{splinemixmeta}(
  object,
  include_smooths = TRUE,
  include_REs = FALSE,
  include_residuals = FALSE,
  type = "outcome",
  ...
)
}
\arguments{
\item{object}{A fitted \code{mixmeta} object returned from \code{\link[=splinemixmeta]{splinemixmeta()}}}

\item{include_smooths}{\code{TRUE} to include the smooth (spline) terms in predictions. Typically one wants these.}

\item{include_REs}{\code{TRUE} to include any additional random effects (beyond the smooths) that
were provided in the \code{random} argument to \code{splinemixmeta}. Omit these if you want to see just the spline predictions.}

\item{include_residuals}{TRUE to include the random effects (one for each datum) for residual
variation not accounted for in the measurement variation (S). Only matters if \code{residual_re = TRUE}
when calling \code{splinemixmeta}, which should typically be the case. Typically one does not want these in predictions.}

\item{type}{Type of predictions. This can be "outcome" or "residual" and will be passed to the \code{type} argument of \code{\link[=blup.splinemixmeta]{blup.splinemixmeta()}}.}

\item{...}{Additional arguments (currently unused)}
}
\value{
A matrix with columns "blup" for the predicted values, "se" for the standard errors of the predictions,
and "vcov" for the variance of the predictions. These are returned from \code{mixmeta::blup()} with \code{vcov=TRUE} and \code{se=TRUE}.
}
\description{
Make predictions from a fitted \code{splinemixmeta} model
}
\details{
This is a convenience function that calls \code{blup} (i.e. \code{blup.splinemixmeta}) without requiring one to know which random-effects "levels"
of the fitted mixmeta object correspond to which parts of the model. Instead one can simply choose whether to include smooths, random effects,
and/or residuals. For more fine-grained control (such as including one spline
term but not another), one can use \code{blup()} directly.
}
