% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_smm.R
\name{predict_smm}
\alias{predict_smm}
\title{Title}
\usage{
predict_smm(
  object,
  include_smooths = TRUE,
  include_REs = FALSE,
  include_residuals = FALSE
)
}
\arguments{
\item{object}{A fitted \code{mixmeta} object returned from \code{splinemixmeta}}

\item{include_smooths}{TRUE to include the smooth (spline) terms in predictions}

\item{include_REs}{TRUE to include any additional random effects (beyond the smooths) that
were provided in the \code{random} argument to \code{splinemixmeta}}

\item{include_residuals}{TRUE to include the random effects (one for each datum) for residual
variation not accounted for in the measurement variation (S). Only matters if \code{residual_re = TRUE}
when calling \code{splinemixmeta}, which should typically be the case.}
}
\value{
A matrix with columns "blup" for the predicted values, "se" for the standard errors of the predictions,
and "vcov" for the variance of the predictions. These are returns from \code{mixmeta::blup()}.
}
\description{
Title
}
