% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm_plot.R
\name{smm_plot}
\alias{smm_plot}
\title{Title}
\usage{
smm_plot(object, xvar, title, xlab, ylab, ylim, color = "blue")
}
\arguments{
\item{object}{mmfit}

\item{xvar}{xvar}

\item{title}{title}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{ylim}{ylim}

\item{color}{color}
}
\value{
ggplot2 object
}
\description{
Title
}
