% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup_smm.R
\name{blup.splinemixmeta}
\alias{blup.splinemixmeta}
\title{Obtain predictions (BLUPs) from a \code{splinemixmeta} object}
\usage{
blup.splinemixmeta(
  object,
  se = FALSE,
  pi = FALSE,
  vcov = FALSE,
  pi.level = 0.95,
  type = "outcome",
  level,
  format,
  aggregate = "stat",
  ...
)
}
\arguments{
\item{object}{An object of class \code{splinemixmeta}, returned by \code{splinemixmeta()}.}

\item{se}{Logical indicating whether to return standard errors of the predictions.}

\item{pi}{Logical indicating whether to return prediction intervals.}

\item{vcov}{Logical indicating whether to return the variance-covariance matrix of the predictions.}

\item{pi.level}{Numeric value between 0 and 1 indicating the confidence level for the prediction intervals. Default is 0.95.}

\item{type}{Character string specifying the type of prediction: "outcome" for predicted outcomes or "residual" for predicted residuals.}

\item{level}{Integer indicating the random effects level for which to obtain predictions. Default is the highest level.}

\item{format}{Character string specifying the format of the output: "matrix" or "list". Default depends on the number of outcomes and whether \code{vcov} is requested.}

\item{aggregate}{Character string specifying how to aggregate the output when multiple outcomes are present}

\item{...}{Additional arguments (currently unused).}
}
\value{
A matrix or list of predicted values (BLUPs), optionally including standard errors, prediction intervals, and variance-covariance matrices.
}
\description{
Obtain predictions (BLUPs) from a \code{splinemixmeta} object
}
\details{
This function is modified from \code{mixmeta::blup.mixmeta} with acknowledgement of the original authors.
It is modified to handle intermediate levels of random effects more carefully.
This is currently a temporary solution that may be replaced in future versions.
}
