% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_smm_smooth.R
\name{make_smm_smooth}
\alias{make_smm_smooth}
\title{Create random effects and fixed effects specifications from an mgcv smooth term}
\usage{
make_smm_smooth(
  smooth,
  data,
  vnames = character(),
  manual_fixed = FALSE,
  envir = parent.frame()
)
}
\arguments{
\item{smooth}{A term created by \code{mgcv::s()}}

\item{data}{Data frame containing the variables used in the smooth}

\item{vnames}{"a vector of names to avoid as dummy variable names in the random effects form", per \code{help(mgcv::smooth2random)}, to which \code{vnames} is passed.}

\item{auto_fixed_effects}{If TRUE, the unpenalized dimensions of the smooth are extracted for use as fixed effects.}
}
\value{
A list with elements \code{basisFxns} (the basis functions for the random effects)
and \code{x_fixed} (the fixed effects design matrix, or NULL if
\code{auto_fixed_effects} is FALSE)
}
\description{
This function is for internal use.
}
