% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smm.R
\name{plot.splinemixmeta}
\alias{plot.splinemixmeta}
\title{Plot results from a univariate \code{splinemixmeta} model}
\usage{
\method{plot}{splinemixmeta}(
  x,
  xvar,
  title,
  xlab,
  ylab,
  ylim,
  linecolor = "blue",
  fillcolor = "blue",
  ...
)
}
\arguments{
\item{x}{object of class \code{mixmeta} returned from \code{splinemixmeta}}

\item{xvar}{xvar name of variable to plot on the horizontal axis.}

\item{title}{title to add to the plot}

\item{xlab}{xlab label for the horizontal axis}

\item{ylab}{ylab label for the vertical axis}

\item{ylim}{ylim limits for the vertical axis}

\item{linecolor}{color for the prediction line}

\item{fillcolor}{color for the prediction confidence band}

\item{...}{additional arguments passed to \code{predict.splinemixmeta()}}
}
\value{
ggplot2 object
}
\description{
Plot results from a univariate \code{splinemixmeta} model
}
\details{
This is not a very general plotting function. It is intended to provide a basic feature
for visualizing a univariate \code{splinemixmeta} fit in a way that:
\itemize{
\item includes fixed effects and spline terms in the predicted values, with 95\% confidence bands
\item Shows the data points with 95\% confidence intervals obtains as +/- 2 times the standard errors (\code{se} or \code{diag(S)}).
\item returns a \code{ggplot2} object that can be further updated.
}
}
