% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinemixmeta.R
\name{splinemixmeta}
\alias{splinemixmeta}
\title{Fit a mixed-effects meta-analysis model with inclusion of a spline term}
\usage{
splinemixmeta(
  smooth = NULL,
  formula,
  se,
  S = se^2,
  manual_fixed = FALSE,
  residual_re = TRUE,
  data,
  random = list(),
  method = "reml",
  bscov = "unstr",
  ...
)
}
\arguments{
\item{smooth}{A smoothing term created by \code{mgcv::s()}}

\item{...}{Additional arguments to be passed to \code{mixmeta::mixmeta()}}
}
\value{
An object of class \code{mixmeta}
}
\description{
Fit a mixed-effects meta-analysis model with inclusion of a spline term
}
